Add the CUDA implementation of the homography algorithm (#7)
The homographer class is refactored to an abstract. Child classes
implement the `transformCoordinates` and `transformImage` methods.

The build is also changed to allow building with native CUDA
architectures and toolkit enabled with a CMake option.

README.md now shows the build status

CMake is configured to dynamically find the build architecture and build a shared library